+++ app/src/components/Context/index.js
@@ -8,17 +8,5 @@ var Context = {
   Auth: {
     State: '',
   },
-
-  Chat: {
-    ChatImageSource: {
-      uri:
-        'https://icons-for-free.com/iconfiles/png/512/add+photo+plus+upload+icon-1320184027779532643.png',
-    },
-    ChatImageValue: null,
-
-    ChatImagePending: null,
-    ChatImageValuePending: null,
-    ChatState: 'invalid',
-  },
 };
 module.exports = Context;
