+++ app/src/screens/Chat/icons.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import {Icon} from '@ui-kitten/components';
+
+module.exports.MenuIcon = (props) => <Icon {...props} name="more-vertical" />;
+module.exports.InfoIcon = (props) => <Icon {...props} name="info" />;
+module.exports.ShareIcon = (props) => <Icon {...props} name="share-outline" />;
+module.exports.LogoutIcon = (props) => <Icon {...props} name="log-out" />;
+module.exports.PhotoIcon = (props) => <Icon {...props} name="image" />;
